# api-docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.QueueTranscriptionRunRequest:
    properties:
      parameters:
        $ref: '#/definitions/models.WhisperXParams'
      profile_id:
        type: string
      profile_name:
        type: string
      reuse_checkpoints:
        type: boolean
    type: object
...
  api.ReorderTranscriptionQueueRequest:
    properties:
      ordered_ids:
        items:
          type: string
        type: array
    type: object
...
  api.TranscriptionQueueResponse:
    properties:
      active_item:
        $ref: '#/definitions/models.TranscriptionQueueItem'
      cleared:
        type: integer
      history:
        items:
          $ref: '#/definitions/models.TranscriptionQueueItem'
        type: array
      items:
        items:
          $ref: '#/definitions/models.TranscriptionQueueItem'
        type: array
      job_id:
        type: string
    type: object
...
  models.AdaptiveExecutionPolicy:
    properties:
      learn:
        type: boolean
      learned_plans:
        items:
          $ref: '#/definitions/models.AdaptivePlanSnapshot'
        type: array
      learning_generation:
        type: integer
      profile_id:
        type: string
      profile_revision:
        type: integer
      snapshot_taken:
        type: boolean
      stages:
        additionalProperties:
          $ref: '#/definitions/models.AdaptiveStagePolicy'
        type: object
    type: object
...
  models.AdaptivePlanSnapshot:
    properties:
      candidate_hash:
        type: string
      learning_generation:
        type: integer
      plan_id:
        type: string
      profile_revision:
        type: integer
      scope_key:
        type: string
      settings:
        $ref: '#/definitions/models.AdaptiveStageSettings'
      stage_key:
        type: string
    type: object
...
  models.AdaptiveStagePolicy:
    properties:
      allow_cpu:
        type: boolean
      allow_shorter_windows:
        type: boolean
      cpu_precision:
        type: string
      device_locked:
        type: boolean
      fixed:
        $ref: '#/definitions/models.AdaptiveStageSettings'
      min_batch_size:
        type: integer
      min_window_seconds:
        type: integer
      overlap_seconds:
        type: number
      window_candidates:
        items:
          type: integer
        type: array
    type: object
  models.AdaptiveStageSettings:
    properties:
      batch_size:
        type: integer
      concurrency:
        type: integer
      device:
        type: string
      overlap_seconds:
        type: number
      precision:
        type: string
      stitching_version:
        type: string
      window_seconds:
        type: number
    type: object
...
  models.TranscriptionQueueItem:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      execution_id:
        type: string
      id:
        type: string
      parameters:
        $ref: '#/definitions/models.WhisperXParams'
      position:
        type: integer
      profile_id:
        type: string
      profile_name:
        type: string
      queued_at:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/models.TranscriptionQueueStatus'
      transcription_job_id:
        type: string
      updated_at:
        type: string
    type: object
  models.TranscriptionQueueStatus:
    enum:
    - queued
    - pending
    - processing
    - completed
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - QueueStatusQueued
    - QueueStatusPending
    - QueueStatusProcessing
    - QueueStatusCompleted
    - QueueStatusFailed
    - QueueStatusCancelled
  models.WhisperXParams:
    properties:
      adaptive_policy:
        $ref: '#/definitions/models.AdaptiveExecutionPolicy'
      align_model:
        description: Alignment settings
        type: string
      api_key:
        description: OpenAI settings
        type: string
      attention_context_left:
        description: NVIDIA model-specific parameters
        type: integer
      attention_context_right:
        type: integer
      audio_chunk_duration:
        type: integer
      batch_size:
        type: integer
      beam_size:
        type: integer
      best_of:
        type: integer
      callback_url:
        description: Webhook settings
        type: string
      chunk_size:
        type: integer
      compression_ratio_threshold:
        type: number
      compute_type:
        type: string
      condition_on_previous_text:
        type: boolean
      device:
        description: Device and computation
        type: string
      device_index:
        type: integer
      diarization_checkpoint:
        type: string
      diarization_device:
        type: string
      diarize:
        description: Diarization settings
        type: boolean
      diarize_model:
        description: 'Options: ''pyannote'', ''nvidia_sortformer'''
        type: string
      fp16:
        type: boolean
      hf_token:
        description: Token and progress
        type: string
      hf_token_source:
        type: string
      highlight_words:
        type: boolean
      initial_prompt:
        type: string
      interpolate_method:
        type: string
      is_multi_track_enabled:
        description: Multi-track transcription settings
        type: boolean
      language:
        type: string
      length_penalty:
        type: number
      logprob_threshold:
        type: number
      max_line_count:
        type: integer
      max_line_width:
        description: Output formatting
        type: integer
      max_new_tokens:
        description: Voxtral settings
        type: integer
      max_speakers:
        type: integer
      min_speakers:
        type: integer
      model:
        description: Model parameters
        type: string
      model_cache_only:
        type: boolean
      model_dir:
        type: string
      model_family:
        description: Model family (whisper or nvidia)
        type: string
      no_align:
        type: boolean
      no_speech_threshold:
        type: number
      nvidia_chunk_duration:
        type: integer
      nvidia_precision:
        type: string
      nvidia_prompt:
        type: string
      nvidia_target_language:
        type: string
      nvidia_timestamps:
        type: boolean
      nvidia_use_chunking:
        type: boolean
      output_format:
        description: Output settings
        type: string
      patience:
        type: number
      print_progress:
        type: boolean
      recovery_mode:
        description: |-
          Empty preserves the historical Auto GPU-to-CPU behavior. Explicit fixed
          and stage_management modes never change device, precision, or context.
        type: string
      return_char_alignments:
        type: boolean
      reuse_checkpoints:
        description: |-
          Nil means exact-compatible checkpoint reuse is enabled. False requests
          new artifacts; it never disables durable persistence.
        type: boolean
      segment_resolution:
        type: string
      speaker_embeddings:
        type: boolean
      suppress_numerals:
        type: boolean
      suppress_tokens:
        type: string
      task:
        description: Task and language
        type: string
      temperature:
        description: Transcription quality settings
        type: number
      temperature_increment_on_fallback:
        type: number
      threads:
        type: integer
      transcription_context:
        description: |-
          Nil context fields inherit the requesting user's defaults when a run is
          queued. A pointer to an empty string explicitly disables that default.
        type: string
      transcription_context_terms:
        type: string
      vad_method:
        description: VAD (Voice Activity Detection) settings
        type: string
      vad_offset:
        type: number
      vad_onset:
        type: number
      verbose:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Audio transcription service using WhisperX
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Scriberr API
  version: "1.0"
paths:
...
  /api/v1/transcription/{id}/queue:
    delete:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TranscriptionQueueResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Clear future queued transcription runs
      tags:
      - transcription
    get:
      description: Return the active per-audio queue item and all future queued runs
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Include terminal queue audit records
        in: query
        name: include_terminal
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TranscriptionQueueResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List sequential transcription runs
      tags:
      - transcription
    post:
      consumes:
      - application/json
      description: Persist one immutable parameter or profile snapshot to run against
        this audio file
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Queued run request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.QueueTranscriptionRunRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.TranscriptionQueueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Queue a sequential transcription run
      tags:
      - transcription
  /api/v1/transcription/{id}/queue/{queue_id}:
    delete:
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Queue item ID
        in: path
        name: queue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TranscriptionQueueResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Cancel a future queued transcription run
      tags:
      - transcription
  /api/v1/transcription/{id}/queue/order:
    put:
      consumes:
      - application/json
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Complete ordered list of waiting queue item IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ReorderTranscriptionQueueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TranscriptionQueueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Reorder future transcription runs
      tags:
      - transcription
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"